+++ openmdao/error_checking/test/test_check_config.py
@@ -4,7 +4,8 @@ from six.moves import range
 
 from openmdao.api import Problem, Group, IndepVarComp, ExecComp
 from openmdao.devtools.testutil import TestLogger
-from openmdao.error_checking.check_config import get_sccs, get_relevant_vars, compute_sys_graph
+from openmdao.utils.graph_utils import all_connected_edges
+from openmdao.error_checking.check_config import get_sccs_topo, get_relevant_vars
 
 
 class MyComp(ExecComp):
@@ -114,7 +115,8 @@ class TestCheckConfig(unittest.TestCase):
         self.assertEqual(warnings[2], "System 'G1.C1' executes out-of-order with respect to its source systems ['G1.C2']")
 
         # test comps_only cycle check
-        sccs = [sorted(s) for s in get_sccs(root, comps_only=True) if len(s) > 1]
+        graph = root.compute_sys_graph(comps_only=True)
+        sccs = [sorted(s) for s in get_sccs_topo(graph) if len(s) > 1]
         self.assertEqual([['C4', 'G1.C1', 'G1.C2']], sccs)
 
     def test_relevance(self):
@@ -147,7 +149,7 @@ class TestCheckConfig(unittest.TestCase):
 
         p.setup(check=False)
 
-        g = compute_sys_graph(p.model, p.model._conn_global_abs_in2out, comps_only=True, save_vars=True)
+        g = p.model.compute_sys_graph(comps_only=True, save_vars=True)
         relevant = get_relevant_vars(g, ['indep1.x', 'indep2.x'], ['C8.y'])
 
         indep1_ins = set(['C3.b', 'C3.c', 'C8.b', 'G1.C1.a', 'G2.C5.a', 'G2.C5.b'])
